feat: move common code to pkg/controller to avoid circular dependecies
