+++ controllers/status/status.go
@@ -19,11 +19,12 @@ limitations under the License.
 package status
 
 import (
-	"github.com/opendatahub-io/opendatahub-operator/v2/apis/components"
 	conditionsv1 "github.com/openshift/custom-resource-status/conditions/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/opendatahub-io/opendatahub-operator/v2/apis/components"
 )
 
 // These constants represent the overall Phase as used by .Status.Phase.
