+++ pkg/controller/types/types.go
@@ -1,0 +1,33 @@
+package types
+
+import (
+	"github.com/go-logr/logr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/opendatahub-io/opendatahub-operator/v2/apis/components"
+	dscv1 "github.com/opendatahub-io/opendatahub-operator/v2/apis/datasciencecluster/v1"
+	dsciv1 "github.com/opendatahub-io/opendatahub-operator/v2/apis/dscinitialization/v1"
+	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
+)
+
+type ResourceObject interface {
+	client.Object
+	components.WithStatus
+}
+
+type WithLogger interface {
+	GetLogger() logr.Logger
+}
+
+type ReconciliationRequest struct {
+	client.Client
+	Instance  client.Object
+	DSC       *dscv1.DataScienceCluster
+	DSCI      *dsciv1.DSCInitialization
+	Platform  cluster.Platform
+	Manifests Manifests
+}
+
+type Manifests struct {
+	Paths map[cluster.Platform]string
+}
